refactor(ims): extract shared session machinery so one implementation serves both
Phase 1 of specs/017-volte-inbound-bridge. Moves eleven transport-agnostic
items out of ims::agent into ims::session, so the host-side cellular service
can use the same registration, renewal and inbound-dispatch code the Wi-Fi
calling path uses rather than a copy (FR-019, SC-008).

Copying would have been faster and would have satisfied those requirements
only in appearance. Two copies of registration and renewal logic drift, and the
drift surfaces on whichever path is tested less -- which, given the Wi-Fi path
is the one in production, would be the new one until it was not.

Deliberately done as a pure move rather than a rewrite, because this touches
the production Wi-Fi path and the live call that would prove no regression
cannot be run from here. Every moved function's body is byte-identical to what
it replaced; the only textual change is rustfmt rewrapping one signature that
grew past the line limit when pub(crate) was added. That property is checked
mechanically rather than asserted, and it means a regression would have to be a
compile error rather than a silent behavioural difference.

What stayed behind is everything that knows about the Wi-Fi path's private link
or its second process: the dispatch loop, invite handling, message relaying and
the veth listeners. What moved is session plumbing, subscription, renewal,
backoff, status mapping and response helpers -- verified beforehand to contain
zero references to veth, the control channel or the second agent.

RegisteredSession and register_session widen from private to crate-visible.
That is the point of the extraction rather than a side effect: the cellular
service needs the type, and a second copy of it would defeat the exercise.

No Wi-Fi test file was touched, and the whole suite passes unmodified -- which
is the strongest evidence available here, but is not the same as a live call.
T006 remains open and is marked as needing the operator, since this environment
has neither the production deployment nor its SIM.
